tests/wpt/web-platform-tests/fullscreen/api/element-request-fullscreen.html

<!DOCTYPE html>
<title>Element#requestFullscreen()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<div id="log"></div>
<script>
async_test(function(t)
{
    t.add_cleanup(() => { document.exitFullscreen() });

    var div = document.querySelector("div");

    document.addEventListener("fullscreenchange", t.step_func(function(event)
    {
        assert_equals(event.target, div, "event.target");
        assert_true(event.bubbles, "event.bubbles");
        assert_false(event.cancelable, "event.cancelable");
        assert_true(event.composed, "event.composed");
        t.done();
    }));

    test_driver.bless("fullscreen", () => div.requestFullscreen())
        .catch(t.step_func(assert_unreached));
});
</script>
